The given commit message seems incorrect and does not align with the diff presented.
Based on the diff, a correct commit message would be:

Add package.json with dependencies

This commit introduces a new package.json file with 'express' and 'nodemon' dependencies. Also, a development script is added which uses 'nodemon' to run the application. This sets up the basic environment and run configuration for the project.