Add custom tooltip functionality for page elements
Implemented a new custom-labelled tooltip functionality to replace the default browser tooltip. The custom label div is appended to the body of the document on page load and on mouse movement, it checks the data-title attribute of the hovered element to determine whether to show or hide the tooltip. The position of the tooltip is calculatively set to avoid window boundaries.