Update default loading styles
The default loading options have been updated to have fullscreen as true and color as white. This change has been made in both the unminimized and minimized versions of the loading JavaScript file. The fullscreen mode will now be used by default, and the loader color will be white by default.