Add comments and update event binding in toggle.js
Detailed explanations have been added for getting elements with the class "toggle" and adding an event listener to them. Furthermore, adjusted the event binding on page load from "document" to "window" in both toggle.js and the minified version, doing so to ensure that toggle event listeners are added once the whole window has loaded.